Rewrite In A More Functional Way
I found a better book for learning functional programming, and this new
approach fits the vibe way better than what I was doing before.

My last approach got stuck on trying to put rules in the type system,
but really functional programming is about using simple blocks to
express complex ideas. This new code does that better by leveraging
basic built-in functions to accomplish what I need.

I'm still not happy with the infinite loop, but I can't think of a
better way to ensure that every element is moved after the shuffle.

The randomness with the shuffle is also not quite right. I might try to
pull the randomness out of the main logic to make it more functional.

I also don't like the escape clause in the loop. That extra "or the
length <= 1" check is adding a bit of error correction logic to a place
it probably doesn't belong.