Start Moving the Shuffle to a Trait Implementation
I made a trait that I can impliment for Vec that lets me define the
"santa_shuffle" and use it similarly to how I use "shuffle". This seems
much closer to the correct solution for what I'm trying to do. I think I
need to work on making things a little more generic though.

Also, the recursion loop has a fatal flaw. Because we're using "self"
for the match comparison, the second round is matching against the
previously shuffled vector, not the original one. I need to capture the
original vector and somehow make sure I'm always matching against it.