Ditch Trait and Make Standalone Function
The trait thing got overcomplicated, so instead I just made a standalone
funciton in lib.rs.

I also added a test that runs through the shuffle 1000 times and makes
sure the end result doesn't have anyone drawing themselves.

I think the next step is to refactor the randomness out of the
santa_shuffle function. The consumer of the function should get to
choose how the randomness is done, but I'm not sure how well that works
with vec.shuffle.